+++ packages/frontend/src/components/error-list/error-list.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { ErrorList } from './error-list';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<ErrorList details={[{ message: 'Error!' }]} />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
