+++ packages/frontend/src/components/field-set/field-set.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { FieldSet } from './field-set';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<FieldSet />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
