+++ packages/frontend/src/components/header/header.test.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { Header } from './header';
+
+jest.mock('react-router-dom', () => ({
+  Link: jest.fn().mockImplementation(() => <a></a>),
+}));
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Header />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
