+++ packages/frontend/src/components/login-page/login-page.test.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { LoginPage } from './login-page';
+
+describe('Search page tests', () => {
+  it('renders without crashing', () => {
+    const div = document.createElement('div');
+    ReactDOM.render(
+      <LoginPage identity={undefined} dispatch={() => {}} />,
+      div
+    );
+    ReactDOM.unmountComponentAtNode(div);
+  });
+});
