+++ packages/frontend/src/components/result-list/result-list.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { ResultList } from './result-list';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<ResultList resultData={[]} />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
