+++ packages/frontend/src/components/social-links/social-links.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { SocialLinks } from './social-links';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<SocialLinks />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
