+++ packages/frontend/src/components/text-input/text-input.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { TextInput } from './text-input';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<TextInput value="" onChange={() => {}} />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
