+++ packages/frontend/src/components/toaster/toaster.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import { Toaster } from './toaster';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Toaster />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
